Abstracted Renderer
Abstracted the renderering stuff out into an interface to make it easier in the future to support different types of rendering. I'd eventually like to add DX12, Opengl 4 and Vulkan renderers, so just doing some ground work to make that easier.

There's still some issues with the graphics stuff leaking, along with the windows stuff leaking (HWND is exposed) which would make linux support a problem in the future.