wonky camera
i was a dumb dumb, that wrote some really dumb coode when i thought about it. mainly, kicking off a thread to render, then joining it because otherwise the function exited and the program exited so the main thread left, and bad times would be had. because i am an idiot.
so i simplify, and just run that loopy function instead of running it on a seperate thread. and everything works.